Bind tmux M-k to emulate iTerm2's clear screen (Cmd+K)
iTerm2's Cmd+K clears all tmux panes and the tmux status bar, without clearing the scrollback buffer, rather than clearing the content & scrollback buffer of the current tmux pane.

Cmd is an unavailable key due to too many hoops to get that special key to work. C-k is an unavailable binding due to tmux-pain-control.

From http://notes.jerzygangi.com/the-only-way-that-actually-works-to-clear-the-screen-in-tmux/