Rely more on Vim buffers, less on tab pages
Vim's tabs are not like tabs in most GUIs. It is not the Vim way to open everything in a new tab. For most tasks, start with buffers. Then, fall back to tabs to use as separate workspaces.

That said, ease the transition, by visualizing open buffers when there's only 1 tab, and recycling existing windows and tabs when opening buffers.