Always show iTerm tab bar
I thought smartly hiding them was saving me screen real estate. However, it required me to interact to see system status, vs. reading to see it. Interaction meant switching tabs temporarily to reveal the tab bar. This was disorienting. I couldn't remember what I was working on or where. With the dark theme tabs, the loss of real estate is a small, unobtrusive price to pay.