Prune gone branches
This is much more aggressive than the `--prune` flag alone. It risks data loss. The reason this level of aggression works for me is, in my SDLC, when the remote deletes a branch, it's authoritative: the work is truly done. Further work atop it should be in a new branch. Most contributors do not work on the same branch. Whether they do or not, they're in communication before deletion.

We rebase a lot, so `--delete` isn't sufficient, either. You need `--force`.

H/T to https://stackoverflow.com/a/38404202/62269.