Always wrap prose, in Prettier
* Apply to all Markdown in this repo

I have Markdown fix-on-save disabled, for working with legacy Markdown formatting on the wild west that is the web. Therefore, I think it is safe to always wrap prose when I _do_ manually format Markdown. I will be more likely to catch when it's an issue, e.g. not preserving line breaks, and I won't manually run the fixer.