Delete global Pylint config
This version of the file fell out of the date and generated tons of warnings of presumably deprecated options. If a project uses Pylint, it should configure Pylint local to the project, not globally in the user's home folder. Besides, I've been relying more on autoformatting, language servers, and type checking, vs. linting.