Compact Starship prompt from 2 lines to 1
This saves space in Warp terminal, where user input is already always on the line after the prompt.

In other terminals, this change puts user input on the same line as the rest of the prompt. This unfortunately means the input prompt is no longer left aligned between commands. A more robust solution would be to vary the Starship line break setting, conditional on the running terminal. In the meantime, to keep the input in other terminals as far left as possible:

* Hide unused package version
* Align right the more variable, often absent command duration